+++ pakyow-core/lib/pakyow/generators/project.rb
@@ -4,24 +4,28 @@ require "bundler"
 require "securerandom"
 
 generator :project do
-  after "generate" do
+  action :bundle do
     Bundler.with_original_env do
       run "bundle install --binstubs", message: "Bundling dependencies"
     end
   end
 
-  after "generate" do
+  action :update_assets do
     Bundler.with_original_env do
       run "bundle exec pakyow assets:update", message: "Updating external assets"
     end
   end
 
-  def generating_locally?
+  # TODO: Can this be private?
+  #
+  private def generating_locally?
     local_pakyow = Gem::Specification.sort_by { |g| [g.name.downcase, g.version] }.group_by(&:name).detect { |k, _| k == "pakyow" }
     !local_pakyow || local_pakyow.last.last.version < Gem::Version.new(Pakyow::VERSION)
   end
 
-  def generate_secret
+  # TODO: Can this be private?
+  #
+  private def generate_secret
     SecureRandom.hex(64)
   end
 end