+++ pakyow-core/lib/pakyow/generators/project/example.rb
@@ -1,17 +1,9 @@
 # frozen_string_literal: true
 
 generator :project, :example, extends: :project do
-  def initialize(source_path)
-    @default_generator = Generators::Project.new(
+  action :generate_default_project do |destination_path, **options|
+    Generators::Project.new(
       ::File.expand_path("../../../generators/project/default", __FILE__)
-    )
-
-    super
-  end
-
-  def generate(*args)
-    @default_generator.generate(*args)
-
-    super
+    ).generate(destination_path, **options)
   end
 end